start fixing compilation errors
